+++ packages/gatsby-theme-aoi-top-layout/gatsby-browser.js
@@ -1,3 +1,6 @@
+/* eslint arrow-body-style: warn */
+/* eslint react/jsx-filename-extension: warn */
+/* eslint import/prefer-default-export: warn */
 import React from 'react';
 import TopLayout from './src/components/TopLayout';