+++ packages/gatsby-theme-aoi-components/CHANGELOG.md
@@ -3,6 +3,18 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+## [1.6.1](https://github.com/cieloazul310/gatsby-aoi/compare/v1.6.0...v1.6.1) (2022-02-05)
+
+
+### Bug Fixes
+
+* Create `tsconfig.eslint.json` to each packages ([752b203](https://github.com/cieloazul310/gatsby-aoi/commit/752b2030108281d36086fc7041dbe965b1046bcb))
+* update peerDependencies versions ([45a63b6](https://github.com/cieloazul310/gatsby-aoi/commit/45a63b62285f8d9fce33d1f840a2cfb58d320dd9))
+
+
+
+
+
 # [1.6.0](https://github.com/cieloazul310/gatsby-aoi/compare/v1.5.0...v1.6.0) (2022-02-05)