+++ packages/gatsby-theme-aoi-utils/CHANGELOG.md
@@ -3,6 +3,22 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+## [1.6.1](https://github.com/cieloazul310/gatsby-aoi/compare/v1.6.0...v1.6.1) (2022-02-05)
+
+
+### Bug Fixes
+
+* Create `tsconfig.eslint.json` to each packages ([752b203](https://github.com/cieloazul310/gatsby-aoi/commit/752b2030108281d36086fc7041dbe965b1046bcb))
+
+
+### Features
+
+* **utils:** fix `useBaseUrl` and create the new `useAbsoluteUrl` hook. ([744353d](https://github.com/cieloazul310/gatsby-aoi/commit/744353d3b9c6e55dfb32a303a6ed90a37bb2b069))
+
+
+
+
+
 # [1.6.0](https://github.com/cieloazul310/gatsby-aoi/compare/v1.5.0...v1.6.0) (2022-02-05)
 
 **Note:** Version bump only for package @cieloazul310/gatsby-theme-aoi-utils